add support for a separation of cluster-internal and public URLs (#1680)
It's more efficient to communicate directly to internal services rather than
through the public domain and load balancer that clients have to go through.

This adds support for specifying a separate `[SERVICE]_PUBLIC_URL` that will be
preferred over the normal URL by the "/api/urls" endpoint.

Also made the `getServerurls()` method more "data driven". Probably less
typescript friendly, which I guess was the main driver for the previous method,
but gives us the benefit of better error messages when a server URL is missing.